% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{ETBBgetweights}
\alias{ETBBgetweights}
\title{Retrieve weights from from Monte-Carlo block samples for the ETBB}
\usage{
ETBBgetweights(n, l, B, seed = NA, c = 0.43)
}
\arguments{
\item{n}{length of observed time series.}

\item{l}{block length.}

\item{B}{the number of Monte-Carlo draws.}

\item{seed}{a random number generating seed (optional).}

\item{c}{a constant controlling the tapering of weight assignments over the blocks.}
}
\value{
a list containing an \code{n} by \code{B} matrix containing the the mass placed on the \code{n} observations by the ETBB empirical distribution in each of the \code{B} Monte-Carlo draws. Another object in the list gives more info.
}
\description{
Retrieve weights from from Monte-Carlo block samples for the ETBB
}
