% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{get.pi.tilde.ETBB}
\alias{get.pi.tilde.ETBB}
\title{Computes expected value of the bootstrap weights assigned to each time point by the ETBB}
\usage{
get.pi.tilde.ETBB(n, l, c = 0.43)
}
\arguments{
\item{n}{the length of the observed time series.}

\item{l}{the block length.}

\item{c}{a number controlling the tapering.}
}
\value{
a vector of length \code{n} containing the expected values of the weights assigned to each time point by the ETBB.
}
\description{
Computes expected value of the bootstrap weights assigned to each time point by the ETBB
}
