% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{get.pi.tilde.MBB}
\alias{get.pi.tilde.MBB}
\title{Computes expected value of the bootstrap weights assigned to each time point by the MBB}
\usage{
get.pi.tilde.MBB(n, l)
}
\arguments{
\item{n}{the length of the observed time series.}

\item{l}{the block length.}
}
\value{
a vector of length \code{n} containing the expected values of the weights assigned to each time point by the MBB.
}
\description{
Computes expected value of the bootstrap weights assigned to each time point by the MBB
}
