% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{IinJ}
\alias{IinJ}
\title{A function asking how many members of I are members of J}
\usage{
IinJ(I, J)
}
\arguments{
\item{I}{a vector}

\item{J}{a vector}
}
\value{
the number of entries in \code{I} which are also in \code{J}.
}
\description{
A function asking how many members of I are members of J
}
