% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{MBBgetweights}
\alias{MBBgetweights}
\title{Retrieve weights from from Monte-Carlo block samples for the MBB}
\usage{
MBBgetweights(n, l, B, seed = NA)
}
\arguments{
\item{n}{length of observed time series.}

\item{l}{block length.}

\item{B}{the number of Monte-Carlo draws.}

\item{seed}{a random number generating seed (optional).}
}
\value{
a list containing an \code{n} by \code{B} matrix containing the the mass placed on the \code{n} observations by the MBB empirical distribution in each of the \code{B} Monte-Carlo draws. Another object in the list gives more info.
}
\description{
Retrieve weights from from Monte-Carlo block samples for the MBB
}
