% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{omega}
\alias{omega}
\title{Function defining the tapered assignment of weights across the ETBB blocks}
\usage{
omega(u, c = 0.43)
}
\arguments{
\item{u}{a real number.}

\item{c}{a constant controlling the tapering.}
}
\value{
the value of the tapering function.
}
\description{
Function defining the tapered assignment of weights across the ETBB blocks
}
