% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{S.tilde}
\alias{S.tilde}
\title{Computes bootstrap expected value of the quantile regression objective function}
\usage{
S.tilde(beta, Y, X, pi.tilde, h, tau)
}
\arguments{
\item{beta}{parameter vector at which to compute the objective function.}

\item{Y}{the vector of response values.}

\item{X}{the design matrix (including a column of ones for the intercept).}

\item{pi.tilde}{weights retrieved from the \code{get.pi.tilde.MBB} of \code{get.pi.tilde.ETBB} functions.}

\item{h}{a scalar bandwidth (bandwidth matrix is \code{h} times identity).}

\item{tau}{the quantile of interest.}
}
\value{
the value of the bootstrap expected value of the quantile regression objective function at the given beta.
This function will be numerically maximized in order to define the SMBB and SETBB versions of the true parameter vector; that is, the maximizer of this function is used as the centering \eqn{\tilde \beta_n} in the pivot \eqn{\sqrt{n}(\hat \beta^*_n - \tilde \beta_n)} rather than the sample estimator.
}
\description{
Computes bootstrap expected value of the quantile regression objective function
}
