% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{D.n.star}
\alias{D.n.star}
\title{Computes D.star from paper}
\usage{
D.n.star(Y, X, beta, tau, pi.star)
}
\arguments{
\item{Y}{the vector of response values.}

\item{X}{the design matrix (including a column of ones for the intercept).}

\item{beta}{parameter vector at which to compute the objective function.}

\item{tau}{the quantile of interest.}

\item{pi.star}{weights retrieved from the \code{BBgetweights} function.}
}
\value{
the value of \eqn{\tilde D_n^*} from the paper.
}
\description{
Computes D.star from paper
}
