% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{getNPPIblksizesQR}
\alias{getNPPIblksizesQR}
\title{Chooses block sizes for MBB, ETBB, SMBB, and SETBB via the NPPI for quantile regression}
\usage{
getNPPIblksizesQR(Y, X, tau, min.in.JAB = 100)
}
\arguments{
\item{Y}{the vector of response values.}

\item{X}{the design matrix (including a column of ones for the intercept).}

\item{tau}{the quantile of interest.}

\item{min.in.JAB}{the minimum number of Monte-Carlos draws desired in each jackknife draw}
}
\value{
a list of the NPPI-selected block sizes for the MBB, SMBB, ETBB, and SETBB.
This function is based on the nonparametric plug-in (NPPI) method discussed in Lahiri (2003), which makes use of the jackknife-after-bootstrap (JAB).
}
\description{
Chooses block sizes for MBB, ETBB, SMBB, and SETBB via the NPPI for quantile regression
}
