% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QregBB.R
\name{QregBB}
\alias{QregBB}
\title{Computes bootstrap expected value of the quantile regression objective function}
\usage{
QregBB(Y, X, tau, l, B = 500, h = NULL, alpha = 0.05)
}
\arguments{
\item{Y}{the vector of response values.}

\item{X}{the design matrix (including a column of ones for the intercept).}

\item{tau}{the quantile of interest.}

\item{l}{block size.}

\item{B}{the number of Monte Carlo bootstrap samples to draw.}

\item{h}{a scalar bandwidth (bandwidth matrix is \code{h} times identity).}

\item{alpha}{a significance level to which the returned confidence intervals will correspond.}
}
\value{
a list containing for the MBB, SMBB, ETBB, and SETBB the set of Monte Carlo draws of the pivot quantity \eqn{\sqrt{n}(\hat \beta^*_n - \tilde \beta_n)}, confidence intervals for each component of \eqn{\beta} corresponding to the specified confidence level, and estimates of the asymptotic covariance matrix of the pivot quantity \eqn{\sqrt{n}(\hat \beta_n - \beta)}.
}
\description{
Computes bootstrap expected value of the quantile regression objective function
}
